+++ lib/issuance.py
@@ -12,9 +12,10 @@ from . import (config, util, exceptions, bitcoin, util)
 
 FORMAT_1 = '>QQ?'
 LENGTH_1 = 8 + 8 + 1
-FORMAT_2 = '>QQ??If42p'
-LENGTH_2 = 8 + 8 + 1 + 1 + 4 + 4 + 42
+FORMAT_2 = '>QQ??If'
+LENGTH_2 = 8 + 8 + 1 + 1 + 4 + 4
 ID = 20
+# NOTE: Pascal strings are used for storing descriptions for backwards‐compatibility.
 
 
 def validate (db, source, destination, asset, quantity, divisible, callable_, call_date, call_price, description, block_index):
@@ -124,10 +125,9 @@ def compose (db, source, transfer_destination, asset, quantity, divisible, calla
 
     asset_id = util.asset_id(asset)
     data = config.PREFIX + struct.pack(config.TXTYPE_FORMAT, ID)
-    data += struct.pack(FORMAT_2, asset_id, quantity, 1 if divisible else 0, 1 if callable_ else 0,
+    curr_format = FORMAT_2 + '{}p'.format(len(description) + 1)
+    data += struct.pack(curr_format, asset_id, quantity, 1 if divisible else 0, 1 if callable_ else 0,
         call_date or 0, call_price or 0.0, description.encode('utf-8'))
-    if len(data) > 80:
-        raise exceptions.IssuanceError('Description is greater than 52 bytes.')
     if transfer_destination:
         destination_outputs = [(transfer_destination, None)]
     else:
@@ -139,8 +139,12 @@ def parse (db, tx, message):
 
     # Unpack message.
     try:
-        if (tx['block_index'] > 50100 or config.TESTNET) and len(message) == LENGTH_2: # Protocol change.
-            asset_id, quantity, divisible, callable_, call_date, call_price, description = struct.unpack(FORMAT_2, message)
+        if (tx['block_index'] > 50100 or config.TESTNET) and len(message) >= LENGTH_2: # Protocol change.
+            curr_format = FORMAT_2 + '{}p'.format(len(message) - LENGTH_2)
+            asset_id, quantity, divisible, callable_, call_date, call_price, description = struct.unpack(curr_format, message)
+            if not (tx['block_index'] >= 125000 or config.TESTNET):  # Protocol change.
+                assert len(description) <= 42
+
             call_price = round(call_price, 6) # TODO: arbitrary
             try:
                 description = description.decode('utf-8')