+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Client Versions ##
+* v9.40.0 (2014-08-19)
+	* allow dividends to be paid to XCP holders (protocol change: 125000)
+	* fixed bug in BTCpay validation
+	* allow null expirations (protocol change: 125000)
+	* assert first block in database is BLOCK_FIRST
+	* arbitrarily long asset descriptions and broadcast texts (protocol change: 125000)
 * v9.39.0 (2014-08-12)
 	* re‐match expired order matches from a new block all at once (protocol change: 95000)
 	* bug in issuance fee (protocol change: 95000/31500)